refactor: :truck: Movi pasta para dentro de fields
