+++ src/features/cadastro-plantio/ui/fields/select-planta/planta-preview-panel.tsx
